+++ src/main/java/com/diagnostico/api/model/BWQuestionnaire.java
@@ -1,7 +1,6 @@
 package com.diagnostico.api.model;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -36,7 +35,7 @@ public class BWQuestionnaire implements Serializable {
 	private UUID id;
 	
 	@Column(name = "total_result")
-	private BigDecimal totalResult;
+	private Integer totalResult;
 	
 	@Enumerated(EnumType.STRING)
 	private QuestionnaireStatus status;
@@ -79,11 +78,11 @@ public class BWQuestionnaire implements Serializable {
 		this.id = id;
 	}
 
-	public BigDecimal getTotalResult() {
+	public Integer getTotalResult() {
 		return totalResult;
 	}
 
-	public void setTotalResult(BigDecimal totalResult) {
+	public void setTotalResult(Integer totalResult) {
 		this.totalResult = totalResult;
 	}
