+++ src/main/resources/db/migration/V07__create_bw_questionnaire.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS bw_questionnaire(
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   status TEXT NOT NULL,
-  total_result REAL NOT NULL,
+  total_result INTEGER NOT NULL,
   created TIMESTAMP DEFAULT current_timestamp,
   modified TIMESTAMP,
   meta jsonb,
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS bw_questionnaire(
 CREATE TABLE IF NOT EXISTS bw_section(
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   section INTEGER NOT NULL,
-  total_result REAL NOT NULL,
+  total_result INTEGER NOT NULL,
   meta jsonb,
   bw_questionnaire_id UUID NOT NULL,
   FOREIGN KEY (bw_questionnaire_id) REFERENCES bw_questionnaire(id)
